Some changes

+++ Lab_1/tritset.cpp
@@ -3,7 +3,7 @@
 TritSet operator &(TritSet& set1, TritSet& set2) {
 	TritSet set3(set1.uint_.size() > set2.uint_.size() ? set1.uint_.size() : set2.uint_.size()); //Create Tritset for result with set1's size or set2's size, which bigger
 
-	for (int i = 0; i < set3.uint_.size(); i++) {
+	for (size_t i = 0; i < set3.uint_.size(); i++) {
 		if (set1.uint_.size() - 1 < i)
 			set3.uint_.at(i) = set2.uint_.at(i);
 		else if (set2.uint_.size() - 1 < i)
@@ -22,7 +22,7 @@ TritSet operator &(TritSet& set1, TritSet& set2) {
 TritSet operator |(TritSet& set1, TritSet& set2) {
 	TritSet set3(set1.uint_.size() > set2.uint_.size() ? set1.uint_.size() : set2.uint_.size()); //Create Tritset for result with set1's size or set2's size, which bigger
 
-	for (int i = 0; i < set3.uint_.size(); i++) {
+	for (size_t i = 0; i < set3.uint_.size(); i++) {
 		if (set1.uint_.size() - 1 < i)
 			set3.uint_.at(i) = set2.uint_.at(i);
 		else if (set2.uint_.size() - 1 < i)
@@ -39,7 +39,7 @@ TritSet operator |(TritSet& set1, TritSet& set2) {
 }
 
 TritSet& operator !(TritSet& set1) {
-	for (int i = 0; i < set1.uint_.size(); i++) {
+	for (size_t i = 0; i < set1.uint_.size(); i++) {
 		if (set1.uint_.at(i) == False)
 			set1.uint_.at(i) = True;
 		else if (set1.uint_.at(i) == True)
